requirements.txt
flask
slack_sdk
requests
python-dotenv
google-generativeai